# transaction: segwit input signing was doing quadratic hashing