As in title, segwit input signing was doing quadratic hashing.
Performance improvements are negligible for typical transactions.
Only looked into this as I was playing on testnet with transactions that had 1000+ inputs.
I think the mental overhead introduced as a result of the changes is also negligible though :)

Some measurements of wall clock time for `Transaction.sign` (with libsecp256k1):
```
  0.11 sec -> 0.08 sec    (  61 p2wpkh-p2sh inputs, 1 output)
  2.48 sec -> 0.75 sec    ( 522 p2wpkh-p2sh inputs, 1 output)
 13.2  sec -> 1.8  sec    (1445 p2wpkh inputs, 1 output)
176.4  sec -> 7.6  sec    (5542 p2wpkh inputs, 1 output)
```